revise: reorganize format functions
